Native 페이지네이션 중복 구현을 통합한다
